+++ apps/fluent-tester/src/FluentTester/TestComponents/Persona/consts.ts
@@ -1,2 +1,10 @@
 export const HOMEPAGE_PERSONA_BUTTON = 'Homepage_Persona_Button';
 export const PERSONA_TESTPAGE = 'Persona_TestPage';
+
+/* E2E Testing Persona 1 */
+export const PERSONA_TEST_COMPONENT = 'Persona_Test_Component'; // A component on each specific test page
+export const PERSONA_ACCESSIBILITY_LABEL = 'E2E testing Persona accessibility label';
+
+/* E2E Testing Persona 2 */
+export const PERSONA_NO_A11Y_LABEL_COMPONENT = 'Persona_No_A11y_label_Component';
+export const PERSONA_TEST_COMPONENT_LABEL = 'Test Persona2 - No Accessibility Label'; // A component on each specific test page