+++ apps/fluent-tester/src/FluentTester/TestComponents/Theme/consts.ts
@@ -1,2 +1,4 @@
 export const HOMEPAGE_THEME_BUTTON = 'Homepage_Theme_Button';
 export const THEME_TESTPAGE = 'Theme_TestPage';
+
+export const THEME_TEST_COMPONENT = 'Theme_Test_Component'; // A component on each specific test page