+++ packages/utils/adapters/src/adapters.ts
@@ -17,6 +17,7 @@ const _viewMask: IFilterMask<IViewProps> = {
   accessibilityState: true,
   accessibilityHint: true,
   accessibilityValue: true,
+  enableFocusRing: true,
   hitSlop: true,
   onLayout: true,
   pointerEvents: true,
@@ -142,6 +143,7 @@ const _textMask: IFilterMask<ITextProps> = {
   onBlurCapture: true,
   onFocus: true,
   onFocusCapture: true,
+  android_hyphenationFrequency: true,
 };
 
 const _imageMask: IFilterMask<IImageProps> = {
@@ -187,6 +189,7 @@ const _imageMask: IFilterMask<IImageProps> = {
   width: true,
   height: true,
   fadeDuration: true,
+  nativeID: true,
 };
 
 export function filterViewProps(propName: string): boolean {