+++ packages/utils/adapters/src/adapters.win32.ts
@@ -27,6 +27,7 @@ const _viewMask: IFilterMask<IViewProps> = {
   accessibilityValue: true,
   animationClass: true,
   cursor: true,
+  enableFocusRing: true,
   hitSlop: true,
   onLayout: true,
   pointerEvents: true,
@@ -156,6 +157,7 @@ const _imageMask: IFilterMask<IImageProps> = {
   width: true,
   height: true,
   fadeDuration: true,
+  nativeID: true,
 };
 
 export function filterViewProps(propName: string): boolean {